Fixed address view and exported it to separate file
