+++ Shoppe/Views/CartView/CartView.swift
@@ -25,38 +25,6 @@ struct CartView: View {
 			color: "White",
 			size: "M",
 			price: 19.99
-		),
-		ProductModel(
-			image: "pinkSkirt",
-			description: "Nike Running Shoes",
-			numberOfItems: 2,
-			color: "Black",
-			size: "42",
-			price: 99.99
-		),
-		ProductModel(
-			image: "pinkSkirt",
-			description: "Cotton T-Shirt",
-			numberOfItems: 1,
-			color: "White",
-			size: "M",
-			price: 19.99
-		),
-		ProductModel(
-			image: "pinkSkirt",
-			description: "Nike Running Shoes",
-			numberOfItems: 2,
-			color: "Black",
-			size: "42",
-			price: 99.99
-		),
-		ProductModel(
-			image: "pinkSkirt",
-			description: "Cotton T-Shirt",
-			numberOfItems: 1,
-			color: "White",
-			size: "M",
-			price: 19.99
 		)
 	]
 
@@ -111,52 +79,6 @@ struct NumberOfItemsView: View {
 			Text("\(number)")
 				.applyFont(.raleway, .bold, 18)
 		}
-	}
-}
-
-struct AddressView: View {
-	@Binding var address: String
-	@State var isAdditingMode: Bool = false
-
-	var body: some View {
-		VStack(alignment: .leading) {
-			Text("Shipping Address")
-				.applyFont(.raleway, .bold, 18)
-
-			HStack(alignment: .bottom) {
-				if isAdditingMode {
-					TextField("Address", text: $address)
-						.textFieldStyle(.roundedBorder)
-						.onSubmit {
-							isAdditingMode = false // Exit editing mode when Enter is pressed
-						}
-				} else {
-					Text(address)
-						.applyFont(.raleway, .regular, 14)
-						.lineLimit(nil) // Allows text to expand dynamically
-
-				}
-
-				ZStack {
-					Circle()
-						.fill(Color.mainBlue)
-						.frame(width: 30, height: 30)
-
-					Image(systemName: "pencil")
-						.foregroundStyle(.white)
-				}
-				.padding(.leading, 10)
-				.onTapGesture {
-					isAdditingMode.toggle()
-				}
-			}
-		}
-		.padding(16) // Ensures space inside the shape
-		.background(
-			RoundedRectangle(cornerRadius: 12)
-				.fill(Color.customGray)
-		)
-		.padding(20) // Outer padding for spacing
 	}
 }
